Rename io_preview to preview
