+++ pyiron_workflow/nodes/macro.py
@@ -14,7 +14,7 @@ from pyiron_snippets.factory import classfactory
 from pyiron_workflow.nodes.composite import Composite
 from pyiron_workflow.has_interface_mixins import HasChannel
 from pyiron_workflow.io import Outputs, Inputs
-from pyiron_workflow.io_preview import ScrapesIO
+from pyiron_workflow.preview import ScrapesIO
 from pyiron_workflow.nodes.static_io import StaticNode
 
 if TYPE_CHECKING:
