+++ pyiron_workflow/nodes/transform.py
@@ -13,7 +13,7 @@ from pandas import DataFrame
 from pyiron_snippets.factory import classfactory
 
 from pyiron_workflow.channels import NOT_DATA
-from pyiron_workflow.io_preview import builds_class_io
+from pyiron_workflow.preview import builds_class_io
 from pyiron_workflow.nodes.static_io import StaticNode
